Update using_spin_functions.rst
Use session wait to specify the wait time in miliseconds, because often a whole second is too long and makes the tests much slower if you have many things that need the spin. Even in cases where one second is not enough, waiting less between retries should save time, as long as the actual process being repeated is not very expensive. I guess ideally the wait would be another argument to spin.