[WIP] Add the side navigation
The main site toctree in guides.rst is hidden because we don't want that
to be rendered on the guides page. Sphinx Wagtail Theme currently only
includes a sidebar nav for visible toctrees - it would be possible to make
this a theme option but for now override their template.

Note that there is now some overlap between the side navigation and the
header links.