Improve alignment / reduce whitespace above page headings
Ideally, I would also slightly reduce the whitespace between the main
navbar and the page content - however this is tricky to do here as
there are no easy / robust CSS selectors we can use. I think this is
an improvement as-is, and better to tackle further with a theme
contribution.