Style the landing page main call-to-action as buttons
With some python hackery to make it work. There are potentially ways to make the
`button` directive able to find document cross-references (rather than hardcoded
links). AI suggested a couple, but these didn't seem to actually work - the
solutions had hallucinated methods that are not present in sphinx and don't
seem to have ever existed.