Reorganise top-level docs structure
Currently, we have `/guides` which is not in the TOC but has useful
content, and `/user_guide` which is a page with just a raw TOC tree.
This pushes things down a level in the TOC and is why we've needed
a separate "guides" link to get back to the page with the actual
content as this does not appear in the TOC at all. It also meant
the "next" link flow between quick start, guides and the user guide
was not in a logical sequence.

Instead, reorganise the TOC structure slightly so that:

* The index is now our main document, so the TOC in the sidebar
  works as expected. The index then automatically renders the
  quick start as the "next" link, and the quick start takes you to
  documentation.
* We remove the old "/user_guide" page which only had the autogenerated
  TOC, instead putting a TOC at the bottom of the documentation page.
  We can add a redirect in RTD for the old `/user_guide.html` URL.
* We therefore don't need the hardcoded "Guides" or "Community" links
  in the top nav, as these replicate the (more logically sequenced) links
  in the main nav.