Move "Extensions" to "Useful Resources" and drop from main nav
If we're only linking to a Github search, I'm not sure this merits
a top navigation link. It feels closer to the unofficial / useful
content links in "Useful Resources" at the moment.

If in future we wanted to list extensions, we could add a separate
page in this section of the docs.